--- src/shared/hooks/useIntersect.tsx
@@ -1,31 +1,0 @@
-import { useCallback, useEffect, useRef } from 'react';
-
-type IntersectHandler = (entry: IntersectionObserverEntry, observer: IntersectionObserver) => void;
-
-const useIntersect = (onIntersect: IntersectHandler, options?: IntersectionObserverInit) => {
-  const ref = useRef<HTMLDivElement>(null);
-  const callback = useCallback(
-    (entries: IntersectionObserverEntry[], observer: IntersectionObserver) => {
-      entries.forEach((entry) => {
-        if (entry.isIntersecting) {
-          onIntersect(entry, observer);
-        }
-      });
-    },
-    [onIntersect]
-  );
-
-  useEffect(() => {
-    if (!ref.current) {
-      return;
-    }
-    const observer = new IntersectionObserver(callback, options);
-    observer.observe(ref.current);
-
-    return () => observer.disconnect();
-  }, [ref, options, callback]);
-
-  return ref;
-};
-
-export default useIntersect;
